test (FakeContext) : 가짜 context를 사용하도록 변경
